pyproject.toml
[project]
name = "azfunc-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-functions>=1.23.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.3",
]
